Build aimet-onnx for Windows/amd64

+++ TrainingExtensions/onnx/src/QcQuantizeInfo.h
@@ -74,6 +74,6 @@ struct QcQuantizeInfo
     bool usePerChannelMode;
     int channelAxis;
     int blockAxis;
-    uint blockSize;
+    size_t blockSize;
     std::string name;
 };
